feat: PermissionType Enum 추가

+++ src/main/java/app/allstackproject/privideo/common/enumStatus/AuthPrincipal.java
@@ -6,7 +6,7 @@ public record AuthPrincipal(
         Long memberId,
         Long orgId,
         Boolean orgIsAdmin,
-        Integer orgPermission,
+        Long orgPermission,
         TokenType tokenType
 ) {
 }
